fix: Widen robot_id to BIGINT in harbor_next.sql to match robot.id
0190_2.16.0_schema.up.sql widened robot.id to bigint (max 9007199254740991)
to escape the int4 range. robot_identity_providers.robot_id and
claim_rules.robot_id recreated that overflow risk for high robot IDs.

+++ make/migrations/postgresql/harbor_next.sql
@@ -43,7 +43,7 @@ CREATE TABLE IF NOT EXISTS identity_providers (
 CREATE TABLE IF NOT EXISTS robot_identity_providers (
     id                   SERIAL PRIMARY KEY,
     identity_provider_id INT NOT NULL REFERENCES identity_providers(id) ON DELETE CASCADE,
-    robot_id             INT NOT NULL REFERENCES robot(id) ON DELETE CASCADE,
+    robot_id             BIGINT NOT NULL REFERENCES robot(id) ON DELETE CASCADE,
     creation_time        TIMESTAMP DEFAULT NOW(),
     UNIQUE (identity_provider_id, robot_id)
 );
@@ -51,7 +51,7 @@ CREATE TABLE IF NOT EXISTS robot_identity_providers (
 CREATE TABLE IF NOT EXISTS claim_rules (
     id                   SERIAL PRIMARY KEY,
     identity_provider_id INT NOT NULL REFERENCES identity_providers(id) ON DELETE CASCADE,
-    robot_id             INT NOT NULL DEFAULT 0,
+    robot_id             BIGINT NOT NULL DEFAULT 0,
     claim_path           TEXT NOT NULL,
     value                TEXT,
     creation_time        TIMESTAMP DEFAULT NOW()
